% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaltools-package.R
\docType{package}
\name{evaltools-package}
\alias{evaltools}
\alias{evaltools-package}
\title{evaltools: Tools for Creating LLM Evaluations with Tool-Based Tasks}
\description{
Create evaluations that test specific data-related abilities of an LLM.
}
\author{
\strong{Maintainer}: Sara Altman \email{sara.altman@posit.co}

}
\keyword{internal}
