% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-loader.R
\name{validate_sample}
\alias{validate_sample}
\title{Validate evaluation sample structure}
\usage{
validate_sample(sample, file, required_fields)
}
\arguments{
\item{sample}{List parsed from YAML.}

\item{file}{Path to the YAML file (for error messages).}

\item{required_fields}{Character vector of required field names in input.}
}
\value{
NULL (called for side effects of aborting on error).
}
\description{
Internal function to validate that a sample has required fields.
}
\keyword{internal}
